refactor: move _filepath_to_python_module tests to loader file
