+++ tests/test_loader.py
@@ -1,10 +1,12 @@
+import os
 import re
 from pathlib import Path
 from unittest.mock import MagicMock, patch
 
 from django.test import override_settings
+from django.conf import settings
 
-from django_components.util.loader import get_component_dirs, get_component_files
+from django_components.util.loader import _filepath_to_python_module, get_component_dirs, get_component_files
 
 from .django_test_setup import setup_test_config
 from .testutils import BaseTestCase
@@ -287,3 +289,50 @@ class ComponentFilesTest(BaseTestCase):
             ],
             [True for _ in range(len(file_paths))]
         )
+
+
+class TestFilepathToPythonModule(BaseTestCase):
+    def test_prepares_path(self):
+        base_path = str(settings.BASE_DIR)
+
+        the_path = os.path.join(base_path, "tests.py")
+        self.assertEqual(
+            _filepath_to_python_module(the_path, base_path, None),
+            "tests",
+        )
+
+        the_path = os.path.join(base_path, "tests/components/relative_file/relative_file.py")
+        self.assertEqual(
+            _filepath_to_python_module(the_path, base_path, None),
+            "tests.components.relative_file.relative_file",
+        )
+
+    def test_handles_nonlinux_paths(self):
+        base_path = str(settings.BASE_DIR).replace("/", "//")
+
+        with patch("os.path.sep", new="//"):
+            the_path = os.path.join(base_path, "tests.py")
+            self.assertEqual(
+                _filepath_to_python_module(the_path, base_path, None),
+                "tests",
+            )
+
+            the_path = os.path.join(base_path, "tests//components//relative_file//relative_file.py")
+            self.assertEqual(
+                _filepath_to_python_module(the_path, base_path, None),
+                "tests.components.relative_file.relative_file",
+            )
+
+        base_path = str(settings.BASE_DIR).replace("//", "\\")
+        with patch("os.path.sep", new="\\"):
+            the_path = os.path.join(base_path, "tests.py")
+            self.assertEqual(
+                _filepath_to_python_module(the_path, base_path, None),
+                "tests",
+            )
+
+            the_path = os.path.join(base_path, "tests\\components\\relative_file\\relative_file.py")
+            self.assertEqual(
+                _filepath_to_python_module(the_path, base_path, None),
+                "tests.components.relative_file.relative_file",
+            )
